requirements.txt
# Revolutionary AGI Forex Trading System - Requirements

# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning & AI
scikit-learn>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.20.0
datasets>=2.0.0
accelerate>=0.20.0
tensorflow>=2.10.0
stable-baselines3>=2.0.0
gymnasium>=1.0.0
shimmy>=2.0.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0

# Web Framework & API
fastapi>=0.100.0
uvicorn>=0.20.0
jinja2>=3.0.0
python-multipart>=0.0.5
aiofiles>=0.8.0

# Database & Storage
sqlalchemy>=1.4.0
alembic>=1.8.0
psycopg2-binary>=2.9.0
redis>=4.0.0

# Data Processing
pyyaml>=6.0
python-dateutil>=2.8.0
pytz>=2022.1

# Networking & HTTP
aiohttp>=3.8.0
requests>=2.28.0
websockets>=10.0
urllib3==1.26.6  # Specific version for compatibility

# Telegram Integration
python-telegram-bot==13.7  # Specific version for compatibility
aiogram>=3.0.0

# Trading & Financial
ccxt>=3.0.0
yfinance>=0.2.0
alpha-vantage>=2.3.0

# Plotting & Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
tqdm>=4.64.0
python-dotenv>=0.19.0
click>=8.0.0
rich>=12.0.0

# Async & Concurrency
asyncio-mqtt>=0.11.0
aioredis>=2.0.0

# Cryptography & Security
cryptography>=3.4.0
bcrypt>=3.2.0
passlib>=1.7.0

# Configuration & Logging
pydantic>=1.10.0
loguru>=0.6.0

# Testing (Optional)
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0

# Development (Optional)
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Quantum Computing Simulation
qiskit>=0.40.0
cirq>=1.0.0

# Additional Scientific Computing
sympy>=1.10.0
networkx>=2.8.0

# Time Series Analysis
statsmodels>=0.13.0
arch>=5.3.0

# Natural Language Processing
nltk>=3.7.0
spacy>=3.4.0
textblob>=0.17.0

# Economic Data
fredapi>=0.5.0
pandas-datareader>=0.10.0

# Blockchain & Crypto
web3>=6.0.0
eth-account>=0.8.0

# IoT & Sensor Data
paho-mqtt>=1.6.0
influxdb-client>=1.30.0

# Satellite & Geospatial Data
rasterio>=1.3.0
geopandas>=0.11.0
folium>=0.12.0
geopy>=2.4.0

# Social Media & News API
praw>=7.0.0
newsapi-python>=0.2.7

# Performance Monitoring
psutil>=5.9.0
memory-profiler>=0.60.0

# Jupyter & Notebooks (Optional)
jupyter>=1.0.0
ipykernel>=6.15.0
notebook>=6.4.0